Consertando erros
Erros de compilação corrigidos
